Update prerequisites.md
The instructions for building cmake include deleting the system version, but then instruct users to run the custom built version from the path where it was built... so the system version could have been kept around.
The instructions could say to `sudo make install` which will set the custom built version AS the new system version, placing it in `/usr/local/bin` which is already on the default search path, thus saving users from having to give a strange path to where cmake was built. The only reason for doing it the other way is if users don't have sudo, as far as I can tell, and if that's the case, they will probably not be able to finish the build for openpose anyway, given all the prerequisites.
It would also save them from the messy situation I got into by following the instructions while pathed to `~/openpose/build` - probably not the place you want an in-place cmake build+installation to reside. :) 

An aside (not much to change in instructions) - although they say "Make sure no error occurred" - there is TONS of output, including numerous error messages, for which the configure or bootstrap process for cmake resolve by downloading missing things. So in fact errors do occur, and scroll past probably unseen, but these errors are normal.